Rename hook to match feature
